# docs/openapi.yaml
openapi: 3.2.0
...
info:
  title: Diagram as Code Rendering Service API
  summary: API thống nhất để render diagram source thành SVG hoặc PNG.
  description: |
    Diagram Gateway là public entry point của nền tảng Diagram as Code.
    Gateway xác thực, giới hạn request, chuẩn hóa options, tra cứu cache và
    chuyển yêu cầu tới Kroki rendering backend.

    Production deployment sử dụng Bearer credential. Local self-hosted profile
    có thể tắt authentication bằng cấu hình rõ ràng.

    Thành công trả raw SVG/PNG. Lỗi trả `application/problem+json` và luôn có
    `requestId` để truy vết.
  version: 1.0.0
  license:
    name: MIT
...
servers:
  - url: http://localhost:9000
    description: Local self-hosted Gateway
  - url: https://{host}
    description: Hosted hoặc private Gateway
    variables:
      host:
        default: diagram.example.com
        description: Tên miền của Diagram Gateway
...
security:
  - bearerAuth: []
paths:
  /api/v1/render:
    post:
      tags: [Rendering]
      operationId: renderDiagram
      summary: Render diagram bằng JSON request
      description: |
        Endpoint chính cho VS Code Extension và GitHub Action. `engine`,
        `format`, `source` và `options` nằm trong JSON body.
      x-srs-requirements: [FR-001, FR-004, FR-005, FR-006, FR-009, FR-010]
      x-local-no-auth-supported: true
      requestBody:
        required: true
        description: Diagram source và tùy chọn render.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenderRequest'
            examples:
              mermaid:
                summary: Mermaid SVG
                dataValue:
                  engine: mermaid
                  format: svg
                  source: |-
                    graph TD
                      A[Source] --> B[Diagram]
                  options:
                    theme: default
                  cache:
                    mode: default
              plantuml:
                summary: PlantUML PNG
                dataValue:
                  engine: plantuml
                  format: png
                  source: |-
                    @startuml
                    Alice -> Bob: Hello
                    @enduml
                  options:
                    theme: plain
              graphviz:
                summary: Graphviz SVG
                dataValue:
                  engine: graphviz
                  format: svg
                  source: digraph G { A -> B }
                  options:
                    layout: dot
              d2:
                summary: D2 SVG không ghi cache
                dataValue:
                  engine: d2
                  format: svg
                  source: client -> gateway -> renderer
                  cache:
                    mode: no-store
      responses:
        '200':
          $ref: '#/components/responses/RenderSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/DiagramSyntaxError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/InvalidRendererOutput'
        '503':
          $ref: '#/components/responses/RendererUnavailable'
        '504':
          $ref: '#/components/responses/RenderTimeout'
  /api/v1/render/{engine}/{format}:
    parameters:
      - $ref: '#/components/parameters/Engine'
      - $ref: '#/components/parameters/OutputFormat'
    post:
      tags: [Rendering]
      operationId: renderDiagramText
      summary: Render diagram từ plain-text body
      description: |
        Render source được gửi trực tiếp với `Content-Type: text/plain`.
        Engine và output format được xác định bởi path.

        Options có thể truyền bằng query parameter `theme` và `option.*`.
        Gateway chỉ chuyển các option nằm trong allowlist của engine.
      x-srs-requirements: [FR-002, FR-004, FR-005, FR-006]
      x-local-no-auth-supported: true
      parameters:
        - $ref: '#/components/parameters/Theme'
        - $ref: '#/components/parameters/CacheMode'
      requestBody:
        required: true
        description: Diagram source UTF-8, tối đa 1 MiB.
        content:
          text/plain:
            schema:
              type: string
              minLength: 1
              maxLength: 1048576
            examples:
              mermaid:
                summary: Mermaid flowchart
                serializedValue: |-
                  graph TD
                    A[Source] --> B[Diagram]
              plantuml:
                summary: PlantUML sequence
                serializedValue: |-
                  @startuml
                  Alice -> Bob: Hello
                  @enduml
              graphviz:
                summary: Graphviz directed graph
                serializedValue: 'digraph G { A -> B }'
              d2:
                summary: D2 connection
                serializedValue: 'client -> gateway -> renderer'
      responses:
        '200':
          $ref: '#/components/responses/RenderSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/DiagramSyntaxError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/InvalidRendererOutput'
        '503':
          $ref: '#/components/responses/RendererUnavailable'
        '504':
          $ref: '#/components/responses/RenderTimeout'
  /api/v1/render/{engine}/{format}/{encodedSource}:
    parameters:
      - $ref: '#/components/parameters/Engine'
      - $ref: '#/components/parameters/OutputFormat'
      - $ref: '#/components/parameters/EncodedSource'
    get:
      tags: [Rendering]
      operationId: renderDiagramEncoded
      summary: Render diagram từ source được encode trong URL
      description: |
        `encodedSource` được tạo bằng cách nén UTF-8 source bằng zlib/deflate,
        sau đó encode Base64 URL-safe không padding. Sau khi decode, source
        không được vượt quá 1 MiB.
      x-srs-requirements: [FR-003, FR-004, FR-005, FR-006]
      x-local-no-auth-supported: true
      parameters:
        - $ref: '#/components/parameters/Theme'
        - $ref: '#/components/parameters/CacheMode'
      responses:
        '200':
          $ref: '#/components/responses/RenderSuccess'
        '400':
          $ref: '#/components/responses/InvalidEncodedSource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '414':
          $ref: '#/components/responses/UriTooLong'
        '422':
          $ref: '#/components/responses/DiagramSyntaxError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/InvalidRendererOutput'
        '503':
          $ref: '#/components/responses/RendererUnavailable'
        '504':
          $ref: '#/components/responses/RenderTimeout'
  /api/v1/engines:
    get:
      tags: [Discovery]
      operationId: listSupportedDiagramTypes
      summary: Liệt kê các loại diagram được hỗ trợ
      description: |
        Trả engine, alias, format, renderer version và availability hiện tại.
        Availability có thể thay đổi khi companion service gặp sự cố.
      x-srs-requirements: [FR-008, FR-033, FR-037]
      security: []
      responses:
        '200':
          description: Danh sách engine và format được hỗ trợ.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineList'
              examples:
                default:
                  summary: Bốn engine MVP
                  dataValue:
                    apiVersion: v1
                    generatedAt: '2026-07-22T05:00:00Z'
                    engines:
                      - id: mermaid
                        aliases: []
                        version: 11.15.0
                        formats: [svg, png]
                        available: true
                      - id: plantuml
                        aliases: [c4plantuml]
                        version: 1.2026.6
                        formats: [svg, png]
                        available: true
                      - id: graphviz
                        aliases: [dot]
                        version: 14.1.3
                        formats: [svg, png]
                        available: true
                      - id: d2
                        aliases: []
                        version: '0.7.1'
                        formats: [svg]
                        available: false
                        unavailableReason: Renderer process is not ready
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceNotReady'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key or GitHub OIDC JWT
      description: |
        API key là opaque bearer secret dạng `dg_<random>`. Gateway chỉ lưu SHA-256
        verifier record; plaintext key nằm trong client secret store. GitHub Action có
        thể gửi GitHub OIDC JWT ngắn hạn trực tiếp; Gateway validate signature, issuer,
        custom audience, expiry và repository/workflow/event/ref policy trước khi tạo
        principal. Local profile có thể tắt authentication.
...
  schemas:
...
    OutputFormat:
      type: string
      enum: [svg, png]
...
    RenderRequest:
      type: object
      additionalProperties: false
      required: [engine, format, source]
      properties:
        engine:
          $ref: '#/components/schemas/DiagramEngine'
        format:
          $ref: '#/components/schemas/OutputFormat'
        source:
          type: string
          minLength: 1
          maxLength: 1048576
          description: Diagram source UTF-8; giới hạn được kiểm tra theo byte sau decode.
        options:
          type: object
          description: |
            Theme và engine-specific options. Gateway chỉ cho phép các option
            nằm trong allowlist của engine.
          properties:
            theme:
              type: string
              minLength: 1
              maxLength: 64
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
        cache:
          type: object
          additionalProperties: false
          properties:
            mode:
              $ref: '#/components/schemas/CacheMode'
...
  responses:
...
    Unauthorized:
      description: Thiếu credential hoặc credential không hợp lệ.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
          example: Bearer realm="diagram-gateway"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unauthenticated:
              dataValue:
                type: /problems/unauthenticated
                title: Chưa xác thực
                status: 401
                code: UNAUTHENTICATED
                message: Bearer credential bị thiếu hoặc không hợp lệ.
                requestId: req_01JABCXYZ
            invalidOidcToken:
              dataValue:
                type: /problems/oidc-token-invalid
                title: GitHub OIDC token không hợp lệ
                status: 401
                code: OIDC_TOKEN_INVALID
                message: GitHub OIDC token signature hoặc standard claims không hợp lệ.
                requestId: req_01JABCXYZ
...